Adds the test_no_cov script for better debugging.
The test command instruments the code for coverage a report. This
makes it hard to figure out where it failed when a test was run. To
circumvent that the test_no_cov command was added which does not
instrument the code for coverage, so errors occur in the code, as
expected.