Distinguishes between internal and external tests.
When not in a production environment the tests will use lib/index_internal.js
Additionally test files with the prefix test_internal_ will be included.

When testing in production the build version of lib/index.js will be
used. Additional test files with _internal_ in the name will not be
tested.

Also cleans up tests a little.