Lets children be named and removes collectionName.
Now multiple children can be added to a model under different names.
When a child is added as a collection it uses the default name Name[],
instead of the collectionName (which no longer exists.)

Fixes all the tests and code for this change.
Fixes all the tests for this change.